refactor map.tsx
-split map.tsx into components for easier development
