+++ frontend/src/pages/Map.tsx
@@ -6,13 +6,6 @@ import mapboxgl, {
   Map as MbMap,
 } from "mapbox-gl";
 
-import type {
-  Feature as GeoFeature,
-  FeatureCollection,
-  GeoJsonProperties,
-  LineString as GeoLineString,
-  Point as GeoPoint,
-} from "geojson";
 import { Feature, Map as OlMap, View } from "ol";
 import { Coordinate } from "ol/coordinate";
 import { click } from "ol/events/condition";
@@ -29,96 +22,21 @@ import { useEffect, useMemo, useRef, useState } from "react";
 import { createBaseTileLayer, type OsmBasemap } from "../maps/baseLayer";
 import { reverseGeocode } from "../maps/geocoder";
 import { useGetConfigQuery, useGetNodesQuery } from "../slices/apiSlice";
-import { INode } from "../types";
 import { convertNodeIdFromIntToHex } from "../utils/convertNodeId";
 
-// --------------------
-// Settings + storage
-// --------------------
-type MapProvider = "osm" | "mapbox";
-
-const LS_KEYS = {
-  provider: "meshinfo.map.provider",
-  mapboxStyle: "meshinfo.map.mapboxStyle",
-  osmBasemap: "meshinfo.map.osmBasemap",
-  recentDays: "meshinfo.map.recentDays",
-  clusterEnabled: "meshinfo.map.clusterEnabled",
-  settingsPanelOpen: "meshinfo.map.settingsPanelOpen",
-};
-
-function readJson<T>(key: string, fallback: T): T {
-  try {
-    const raw = localStorage.getItem(key);
-    if (!raw) return fallback;
-    return JSON.parse(raw) as T;
-  } catch {
-    return fallback;
-  }
-}
-
-function writeJson<T>(key: string, value: T) {
-  try {
-    localStorage.setItem(key, JSON.stringify(value));
-  } catch (err) {
-    console.warn("Failed to persist map setting to localStorage", { key, error: err });
-  }
-}
-
-function toMapboxStyleUrl(stylePath: string): string {
-  // Accept:
-  // - "mapbox://styles/..."
-  // - "mapbox/streets-v12" or "user/styleid"
-  if (stylePath.startsWith("mapbox://")) return stylePath;
-  return `mapbox://styles/${stylePath}`;
-}
-
-// --------------------
-// Utilities
-// --------------------
-function escapeHtml(text: string): string {
-  const div = document.createElement('div');
-  div.textContent = text;
-  return div.innerHTML;
-}
-
-function calculateGeodesicDistance(lat1: number, lon1: number, lat2: number, lon2: number): number {
-  const R = 6371; // Earth's radius in kilometers
-  const dLat = (lat2 - lat1) * Math.PI / 180;
-  const dLon = (lon2 - lon1) * Math.PI / 180;
-  const a = Math.sin(dLat/2) * Math.sin(dLat/2) +
-    Math.cos(lat1 * Math.PI / 180) * Math.cos(lat2 * Math.PI / 180) * 
-    Math.sin(dLon/2) * Math.sin(dLon/2);
-  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1-a));
-  return R * c;
-}
-
-// --------------------
-// Types
-// --------------------
-type IMapNode = INode & {
-  online: boolean;
-  map_position?: Coordinate; // [lon, lat]
-  neighbors?: {
-    id: string;
-    snr: number;
-    distance: number;
-  }[];
-};
-
-// OL Feature properties for click handling
-type IFeatureNode = {
-  id: string;
-  shortname?: string;
-  longname?: string;
-  last_seen?: string;
-  position: Coordinate; // [lon, lat]
-  online: boolean;
-  neighbors?: {
-    id: string;
-    snr: number;
-    distance: number;
-  }[];
-};
+import type { IFeatureNode, IMapNode, MapProvider, NodeLike } from "./map/types";
+import { LS_KEYS, readJson, toMapboxStyleUrl, writeJson } from "./map/storage";
+import {
+  applyMapboxClusterVisibility,
+  buildNodesGeoJSON,
+  bumpOlRender,
+  computeRecentNodes,
+  emptyLineFeatureCollection,
+} from "./map/utils";
+import { clearDetailsPanel, getDetailsDom, setDetailsPanelContent } from "./map/detailsDom";
+import { buildMapboxLinkFeatureCollection, buildNodeDetailsHtml } from "./map/detailsHtml";
+import { MapDetailsPanel } from "./map/MapDetailsPanel";
+import { MapSettingsPanel } from "./map/MapSettingsPanel";
 
 // --------------------
 // OpenLayers styles
@@ -147,98 +65,6 @@ const onlineStyle = new Style({
   }),
 });
 
-// --------------------
-// Helpers
-// --------------------
-function computeRecentNodes(
-  nodes: Record<string, IMapNode>,
-  recentDays: number
-) {
-  const recentCutoff = Date.now() - recentDays * 24 * 60 * 60 * 1000;
-
-  return Object.entries(nodes).filter(([_, node]) => {
-    if (node.online) return true;
-    if (!node.last_seen) return false;
-
-    const lastSeenMs = new Date(node.last_seen).getTime();
-    if (Number.isNaN(lastSeenMs)) return false;
-
-    return lastSeenMs > recentCutoff;
-  });
-}
-
-function bumpOlRender(map: OlMap) {
-  map.updateSize();
-  map.renderSync();
-
-  requestAnimationFrame(() => {
-    map.updateSize();
-    map.renderSync();
-  });
-
-  // One more delayed bump catches late layout/font/sidebar shifts.
-  window.setTimeout(() => {
-    map.updateSize();
-    map.renderSync();
-  }, 200);
-}
-
-
-function buildNodesGeoJSON(
-  nodes: Record<string, IMapNode>,
-  recentDays: number
-): FeatureCollection<GeoPoint, GeoJsonProperties> {
-  const recentNodeEntries = computeRecentNodes(nodes, recentDays);
-
-  const features: GeoFeature<GeoPoint, GeoJsonProperties>[] = [];
-
-  for (const [id, node] of recentNodeEntries) {
-    if (!node.map_position) continue;
-
-    features.push({
-      type: "Feature",
-      id, // important for feature-state selection
-      properties: {
-        id,
-        shortname: node.shortname ?? "",
-        longname: node.longname ?? "",
-        last_seen: node.last_seen ?? "",
-        online: Boolean(node.online),
-      },
-      geometry: {
-        type: "Point",
-        coordinates: [node.map_position[0], node.map_position[1]],
-      },
-    });
-  }
-
-  return { type: "FeatureCollection", features };
-}
-
-function emptyLineFeatureCollection(): FeatureCollection<
-  GeoLineString,
-  GeoJsonProperties
-> {
-  return { type: "FeatureCollection", features: [] };
-}
-
-function applyMapboxClusterVisibility(map: MbMap, enabled: boolean): void {
-  const set = (layerId: string, visible: boolean) => {
-    if (!map.getLayer(layerId)) return;
-    map.setLayoutProperty(layerId, "visibility", visible ? "visible" : "none");
-  };
-
-  // clustered set
-  set("clusters", enabled);
-  set("cluster-count", enabled);
-  set("unclustered-nodes", enabled);
-  set("unclustered-labels", enabled);
-
-  // plain set
-  set("plain-nodes", !enabled);
-  set("plain-labels", !enabled);
-}
-
 export function Map() {
   const mapRef = useRef<HTMLDivElement>(null);
 
@@ -248,9 +74,7 @@ export function Map() {
 
   // OL map state (OSM path)
   const [olMap, setOlMap] = useState<OlMap>();
-  const olBaseLayerRef = useRef<ReturnType<typeof createBaseTileLayer> | null>(
-    null
-  );
+  const olBaseLayerRef = useRef<ReturnType<typeof createBaseTileLayer> | null>(null);
   const olNodesSourceRef = useRef<VectorSource<Feature<Point>> | null>(null);
 
   // Mapbox refs (Mapbox path)
@@ -303,7 +127,7 @@ export function Map() {
   const [settingsPanelOpen, setSettingsPanelOpen] = useState<boolean>(() => {
     const stored = readJson<boolean | null>(LS_KEYS.settingsPanelOpen, null);
     // Default to false on mobile, true on desktop
-    return stored ?? (typeof window !== 'undefined' && window.innerWidth >= 1024);
+    return stored ?? (typeof window !== "undefined" && window.innerWidth >= 1024);
   });
 
   // persist settings
@@ -311,14 +135,8 @@ export function Map() {
   useEffect(() => writeJson(LS_KEYS.mapboxStyle, mapboxStyle), [mapboxStyle]);
   useEffect(() => writeJson(LS_KEYS.osmBasemap, osmBasemap), [osmBasemap]);
   useEffect(() => writeJson(LS_KEYS.recentDays, recentDays), [recentDays]);
-  useEffect(
-    () => writeJson(LS_KEYS.clusterEnabled, clusterEnabled),
-    [clusterEnabled]
-  );
-  useEffect(
-    () => writeJson(LS_KEYS.settingsPanelOpen, settingsPanelOpen),
-    [settingsPanelOpen]
-  );
+  useEffect(() => writeJson(LS_KEYS.clusterEnabled, clusterEnabled), [clusterEnabled]);
+  useEffect(() => writeJson(LS_KEYS.settingsPanelOpen, settingsPanelOpen), [settingsPanelOpen]);
 
   // If token disappears / not configured, force provider to osm
   useEffect(() => {
@@ -421,50 +239,21 @@ export function Map() {
     clusterEnabledRef.current = clusterEnabled;
   }, [clusterEnabled]);
 
-  // ----------------------------
-  // Shared DOM refs for the side panel
-  // ----------------------------
-  function getDetailsDom() {
-    const nodePanel = document.getElementById("details");
-    const nodeTitle = document.getElementById("details-title");
-    const nodeSubtitle = document.getElementById("details-subtitle");
-    const nodeContent = document.getElementById("details-content");
-
-    return { nodePanel, nodeTitle, nodeSubtitle, nodeContent };
-  }
-
-  function clearDetailsPanel() {
-    const { nodePanel, nodeTitle, nodeSubtitle, nodeContent } = getDetailsDom();
-    if (!nodePanel || !nodeTitle || !nodeSubtitle || !nodeContent) return;
-
-    nodeTitle.textContent = "";
-    nodeSubtitle.textContent = "";
-    nodeContent.innerHTML = "";
-
-    nodePanel.classList.add("hidden");
-  }
-
   function clearMapboxSelectionAndOverlays() {
-  const map = mbMapRef.current;
-  const selectedId = mbSelectedIdRef.current;
+    const map = mbMapRef.current;
+    const selectedId = mbSelectedIdRef.current;
 
     if (map && selectedId) {
       // Clear selection ring (feature-state) for both sources (clustered + plain)
       try {
         if (map.getSource("nodes_clustered")) {
-          map.setFeatureState(
-            { source: "nodes_clustered", id: selectedId },
-            { selected: false }
-          );
+          map.setFeatureState({ source: "nodes_clustered", id: selectedId }, { selected: false });
         }
       } catch {}
 
       try {
         if (map.getSource("nodes_plain")) {
-          map.setFeatureState(
-            { source: "nodes_plain", id: selectedId },
-            { selected: false }
-          );
+          map.setFeatureState({ source: "nodes_plain", id: selectedId }, { selected: false });
         }
       } catch {}
     }
@@ -499,7 +288,6 @@ export function Map() {
 
     if (provider !== "osm" && olMap) {
       olMap.setTarget(undefined);
-
       if (mapRef.current) mapRef.current.innerHTML = "";
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -530,12 +318,29 @@ export function Map() {
         }
       : defaultPosition;
 
-    const savedCenter = JSON.parse(localStorage.getItem("savedCenter") ?? "[]");
+    // Safer savedCenter parsing (avoid crash if LS is corrupt)
+    let savedCenter: unknown = [];
+    try {
+      savedCenter = JSON.parse(localStorage.getItem("savedCenter") ?? "[]");
+    } catch {
+      savedCenter = [];
+    }
+    const saved = Array.isArray(savedCenter) ? savedCenter : [];
+    const savedLon = typeof saved[0] === "number" ? saved[0] : undefined;
+    const savedLat = typeof saved[1] === "number" ? saved[1] : undefined;
+
     const initialCenter: [number, number] = [
-      savedCenter[0] ?? centerPos.longitude,
-      savedCenter[1] ?? centerPos.latitude,
+      savedLon ?? centerPos.longitude,
+      savedLat ?? centerPos.latitude,
     ];
-    const initialZoom = JSON.parse(localStorage.getItem("savedZoom") ?? "9.5");
+
+    let initialZoom = 9.5;
+    try {
+      const z = JSON.parse(localStorage.getItem("savedZoom") ?? "9.5");
+      if (typeof z === "number" && Number.isFinite(z)) initialZoom = z;
+    } catch {
+      // ignore
+    }
 
     const styleUrl = toMapboxStyleUrl(mapboxStyle);
     mbCurrentStyleUrlRef.current = styleUrl;
@@ -820,137 +625,31 @@ export function Map() {
 
         const displayName = await reverseGeocode(node.map_position[0], node.map_position[1]);
 
-        let panel =
-          `<b>${escapeHtml(node.longname ?? "")}</b><br/>${escapeHtml(node.shortname ?? "")} / ${escapeHtml(id)}<br/><br/>` +
-          `<b>Position</b><br/>${escapeHtml(node.map_position.toString())}<br/><br/>` +
-          `<b>Location</b><br/>${escapeHtml(displayName || "Unknown")}<br/><br/>` +
-          `<b>Status</b><br/>${node.online ? "Online" : "Offline"}<br/><br/>` +
-          `<b>Last Seen</b><br/>${escapeHtml(node.last_seen ?? "")}<br/><br/>`;
+        const nodeLike: NodeLike = {
+          id,
+          shortname: node.shortname,
+          longname: node.longname,
+          last_seen: node.last_seen,
+          online: Boolean(node.online),
+          position: node.map_position,
+          neighbors: node.neighbors,
+        };
 
-        panel += "<b>Neighbors Heard</b><br/>";
-        if ((node.neighbors?.length ?? 0) === 0) {
-          panel += "None";
-        } else {
-          panel +=
-            "<table border=1 cellpadding=2 cellspacing=0 width=100% class='border border-gray-300'>";
-          panel +=
-            "<tr><th width=33% align=left>Node</th><th width=33% align=center>SNR</th><th width=33% align=right>Distance</th></tr>";
-
-          panel += (node.neighbors ?? [])
-            .map((neighbor) => {
-              const nnode = liveNodes[neighbor.id];
-              if (!nnode) {
-                return `<tr><td class="text-gray-600">UNK</td><td align=center>${neighbor.snr}</td><td></td></tr>`;
-              }
-
-              let distance;
-              if (nnode.map_position) {
-                distance = calculateGeodesicDistance(
-                  node.map_position![1], node.map_position![0],
-                  nnode.map_position[1], nnode.map_position[0]
-                );
-              }
-
-              return `<tr><td align=left>${escapeHtml(nnode.shortname ?? "")}</td><td align=center>${neighbor.snr}</td><td align=right>${
-                distance ? distance.toFixed(2) : "unk"
-              } km</td></tr>`;
-            })
-            .join("");
-
-          panel += "</table>";
-        }
-        panel += "<br/><br/>";
-
-        panel += "<b>Heard By Neighbors</b><br/>";
-        const heardBy = Object.keys(liveNodes).filter((nid) =>
-          liveNodes[nid].neighbors?.some((neighbor) => neighbor.id === id)
-        );
-
-        if (heardBy.length === 0) {
-          panel += "None<br/>";
-        } else {
-          panel +=
-            "<table border=1 cellpadding=2 cellspacing=0 width=100% class='border border-gray-300'>";
-          panel +=
-            "<tr><th width=33% align=left>Node</th><th width=33% align=center>SNR</th><th width=33% align=right>Distance</th></tr>";
-
-          panel += heardBy
-            .map((nid) => {
-              const nnode = liveNodes[nid];
-              const neighbor = nnode?.neighbors?.find((n) => n.id === id);
-
-              if (!nnode) {
-                return `<tr><td class="text-gray-600">UNK</td><td align=center>${neighbor?.snr}</td><td></td></tr>`;
-              }
-
-              let distance;
-              if (nnode.map_position) {
-                distance = calculateGeodesicDistance(
-                  node.map_position![1], node.map_position![0],
-                  nnode.map_position[1], nnode.map_position[0]
-                );
-              }
-
-              return `<tr><td align=left>${escapeHtml(nnode.shortname ?? "")}</td><td align=center>${neighbor?.snr}</td><td align=right>${
-                distance ? distance.toFixed(2) : "unk"
-              } km</td></tr>`;
-            })
-            .join("");
-
-          panel += "</table>";
-        }
-
-        panel += "<br/><br/>";
-
-        panel += "<b>Elsewhere</b><br/>";
-        const nodeId = parseInt(id, 16);
-        panel += `<a class="dark:text-indigo-400 dark:visited:text-indigo-400 dark:hover:text-indigo-500" href="https://meshview.armooo.net/packet_list/${nodeId}" target="_blank">Armooo's MeshView</a><br/>`;
-        panel += `<a class="dark:text-indigo-400 dark:visited:text-indigo-400 dark:hover:text-indigo-500" href="https://app.bayme.sh/node/${encodeURIComponent(id)}" target="_blank">Bay Mesh Explorer</a><br/>`;
-        panel += `<a class="dark:text-indigo-400 dark:visited:text-indigo-400 dark:hover:text-indigo-500" href="https://meshtastic.liamcottle.net/?node_id=${nodeId}" target="_blank">Liam's Map</a><br/>`;
-        panel += `<a class="dark:text-indigo-400 dark:visited:text-indigo-400 dark:hover:text-indigo-500" href="https://meshmap.net/#${nodeId}" target="_blank">MeshMap</a><br/>`;
-
-        const { nodePanel, nodeTitle, nodeSubtitle, nodeContent } = getDetailsDom();
-        if (nodePanel && nodeTitle && nodeSubtitle && nodeContent) {
-          nodeTitle.textContent = node.longname ?? "";
-          nodeSubtitle.textContent = node.shortname ?? "";
-          nodeContent.innerHTML = panel;
-          nodePanel.classList.remove("hidden");
-        }
-
-        // Draw links
-        const linkFeatures: GeoFeature<GeoLineString, GeoJsonProperties>[] = [];
-
-        const neighborSet = new Set((node.neighbors ?? []).map((n) => n.id));
-        const heardBySet = new Set(heardBy);
-        const union = new Set<string>([...neighborSet, ...heardBySet]);
-
-        union.forEach((otherId) => {
-          const other = liveNodes[otherId];
-          if (!other?.map_position) return;
-
-          const isNeighbor = neighborSet.has(otherId);
-          const isHeardBy = heardBySet.has(otherId);
-          const kind =
-            isNeighbor && isHeardBy ? "both" : isNeighbor ? "neighbor" : "heard_by";
-
-          linkFeatures.push({
-            type: "Feature",
-            properties: { kind },
-            geometry: {
-              type: "LineString",
-              coordinates: [
-                [node.map_position![0], node.map_position![1]],
-                [other.map_position[0], other.map_position[1]],
-              ],
-            },
-          });
+        const { html, heardBy } = buildNodeDetailsHtml({
+          node: nodeLike,
+          liveNodes,
+          displayName,
         });
 
+        setDetailsPanelContent({
+          title: node.longname ?? "",
+          subtitle: node.shortname ?? "",
+          html,
+        });
+
+        // Draw links
         const linksSource = map.getSource("links") as MbGeoJSONSource | undefined;
-        linksSource?.setData({
-          type: "FeatureCollection",
-          features: linkFeatures,
-        } as FeatureCollection<GeoLineString, GeoJsonProperties>);
+        linksSource?.setData(buildMapboxLinkFeatureCollection({ node: nodeLike, liveNodes, heardBy }));
       };
 
       // Cursor behaviors (both render modes)
@@ -1007,8 +706,9 @@ export function Map() {
       // Clicking empty space clears
       map.on("click", (e) => {
         const hitNode =
-          map.queryRenderedFeatures(e.point, { layers: ["unclustered-nodes", "plain-nodes", "unclustered-labels", "plain-labels"], })
-            .length > 0;
+          map.queryRenderedFeatures(e.point, {
+            layers: ["unclustered-nodes", "plain-nodes", "unclustered-labels", "plain-labels"],
+          }).length > 0;
         const hitCluster =
           map.queryRenderedFeatures(e.point, { layers: ["clusters"] }).length > 0;
         if (hitNode || hitCluster) return;
@@ -1052,8 +752,7 @@ export function Map() {
       linksSource?.setData(emptyLineFeatureCollection());
 
       map.setStyle(desired);
-    } catch {
-    }
+    } catch {}
   }, [mapboxStyle, provider, hasMapbox]);
 
   // Mapbox: cluster toggle
@@ -1082,18 +781,15 @@ export function Map() {
     // If selected node disappears, clear selection + links/panel
     const selectedId = mbSelectedIdRef.current;
     if (selectedId) {
-      const stillExists = data.features.some(
-        (f) => (f.properties?.id as string | undefined) === selectedId
-      );
+      const stillExists = data.features.some((f) => (f.properties?.id as string | undefined) === selectedId);
       if (!stillExists) {
         try {
           map.setFeatureState({ source: "nodes_clustered", id: selectedId }, { selected: false });
-        } catch {
-        }
+        } catch {}
         try {
           map.setFeatureState({ source: "nodes_plain", id: selectedId }, { selected: false });
-        } catch {
-        }
+        } catch {}
+
         mbSelectedIdRef.current = null;
 
         const linksSource = map.getSource("links") as MbGeoJSONSource | undefined;
@@ -1131,9 +827,7 @@ export function Map() {
 
     // Prefer serverNode if it has a position, otherwise fall back to any node with a position
     const fallbackNodeWithPos =
-      serverNode?.map_position
-        ? serverNode
-        : Object.values(nodes).find((n) => n.map_position);
+      serverNode?.map_position ? serverNode : Object.values(nodes).find((n) => n.map_position);
 
     const centerPos = fallbackNodeWithPos?.map_position
       ? {
@@ -1154,10 +848,7 @@ export function Map() {
     const savedLon = typeof saved[0] === "number" ? saved[0] : undefined;
     const savedLat = typeof saved[1] === "number" ? saved[1] : undefined;
 
-    const initialCenter = fromLonLat([
-      savedLon ?? centerPos.longitude,
-      savedLat ?? centerPos.latitude,
-    ]);
+    const initialCenter = fromLonLat([savedLon ?? centerPos.longitude, savedLat ?? centerPos.latitude]);
 
     let initialZoom = 9.5;
     try {
@@ -1270,126 +961,54 @@ export function Map() {
 
       const displayName = await reverseGeocode(node.position[0], node.position[1]);
 
-      let panel =
-        `<b>${escapeHtml(node.longname ?? "")}</b><br/>${escapeHtml(node.shortname ?? "")} / ${escapeHtml(node.id)}<br/><br/>` +
-        `<b>Position</b><br/>${escapeHtml(node.position.toString())}<br/><br/>` +
-        `<b>Location</b><br/>${escapeHtml(displayName || "Unknown")}<br/><br/>` +
-        `<b>Status</b><br/>${node.online ? "Online" : "Offline"}<br/><br/>` +
-        `<b>Last Seen</b><br/>${escapeHtml(node.last_seen ?? "")}<br/><br/>`;
+      const nodeLike: NodeLike = {
+        id: node.id,
+        shortname: node.shortname,
+        longname: node.longname,
+        last_seen: node.last_seen,
+        online: Boolean(node.online),
+        position: node.position,
+        neighbors: node.neighbors,
+      };
 
-      panel += "<b>Neighbors Heard</b><br/>";
-      if ((node.neighbors?.length ?? 0) === 0) {
-        panel += "None";
-      } else {
-        panel +=
-          "<table border=1 cellpadding=2 cellspacing=0 width=100% class='border border-gray-300'>";
-        panel +=
-          "<tr><th width=33% align=left>Node</th><th width=33% align=center>SNR</th><th width=33% align=right>Distance</th></tr>";
+      const { html } = buildNodeDetailsHtml({
+        node: nodeLike,
+        liveNodes: nodes,
+        displayName,
+      });
 
-        panel += (node.neighbors ?? [])
-          .map((neighbor) => {
-            const nnode = nodes[neighbor.id];
-            if (!nnode) {
-              return `<tr><td class="text-gray-600">UNK</td><td align=center>${neighbor.snr}</td><td></td></tr>`;
-            }
+      setDetailsPanelContent({
+        title: node.longname ?? "",
+        subtitle: node.shortname ?? "",
+        html,
+      });
 
-            let distance;
-            if (nnode.map_position) {
-              distance = calculateGeodesicDistance(
-                node.position[1], node.position[0],
-                nnode.map_position[1], nnode.map_position[0]
-              );
-            }
+      // Draw neighbor lines (OpenLayers map overlay)
+      node.neighbors?.forEach((neighbor) => {
+        const nnode = nodes[neighbor.id];
+        if (!nnode?.map_position) return;
 
-            return `<tr><td align=left>${escapeHtml(nnode.shortname ?? "")}</td><td align=center>${neighbor.snr}</td><td align=right>${
-              distance ? distance.toFixed(2) : "unk"
-            } km</td></tr>`;
-          })
-          .join("");
+        const points: Coordinate[] = [node.position, nnode.map_position];
 
-        panel += "</table>";
+        for (let i = 0; i < points.length; i++) {
+          points[i] = transform(points[i], "EPSG:4326", "EPSG:3857");
+        }
 
-        node.neighbors?.forEach((neighbor) => {
-          const nnode = nodes[neighbor.id];
-          if (!nnode?.map_position) return;
+        const featureLine = new Feature({ geometry: new LineString(points) });
 
-          const points: Coordinate[] = [node.position, nnode.map_position];
+        const vectorLine = new Vector({});
+        vectorLine.addFeature(featureLine);
 
-          for (let i = 0; i < points.length; i++) {
-            points[i] = transform(points[i], "EPSG:4326", "EPSG:3857");
-          }
-
-          const featureLine = new Feature({ geometry: new LineString(points) });
-
-          const vectorLine = new Vector({});
-          vectorLine.addFeature(featureLine);
-
-          const vectorLineLayer = new VectorLayer({
-            source: vectorLine,
-            style: new Style({
-              fill: new Fill({ color: "#66FF66" }),
-              stroke: new Stroke({ color: "#66FF66", width: 4 }),
-            }),
-          });
-          neighborLayers.push(vectorLineLayer);
-          map.addLayer(vectorLineLayer);
+        const vectorLineLayer = new VectorLayer({
+          source: vectorLine,
+          style: new Style({
+            fill: new Fill({ color: "#66FF66" }),
+            stroke: new Stroke({ color: "#66FF66", width: 4 }),
+          }),
         });
-      }
-
-      panel += "<br/><br/>";
-
-      panel += "<b>Heard By Neighbors</b><br/>";
-      const heardBy = Object.keys(nodes).filter((nid) =>
-        nodes[nid].neighbors?.some((neighbor) => neighbor.id === node.id)
-      );
-
-      if (heardBy.length === 0) {
-        panel += "None<br/>";
-      } else {
-        panel +=
-          "<table border=1 cellpadding=2 cellspacing=0 width=100% class='border border-gray-300'>";
-        panel +=
-          "<tr><th width=33% align=left>Node</th><th width=33% align=center>SNR</th><th width=33% align=right>Distance</th></tr>";
-
-        panel += heardBy
-          .map((nid) => {
-            const nnode = nodes[nid];
-            const neighbor = nnode?.neighbors?.find((n) => n.id === node.id);
-
-            if (!nnode) {
-              return `<tr><td class="text-gray-600">UNK</td><td align=center>${neighbor?.snr}</td><td></td></tr>`;
-            }
-
-            let distance;
-            if (nnode.map_position) {
-              distance = calculateGeodesicDistance(
-                node.position[1], node.position[0],
-                nnode.map_position[1], nnode.map_position[0]
-              );
-            }
-
-            return `<tr><td align=left>${escapeHtml(nnode.shortname ?? "")}</td><td align=center>${neighbor?.snr}</td><td align=right>${
-              distance ? distance.toFixed(2) : "unk"
-            } km</td></tr>`;
-          })
-          .join("");
-
-        panel += "</table>";
-      }
-
-      panel += "<br/><br/>";
-
-      panel += "<b>Elsewhere</b><br/>";
-      const nodeId = parseInt(node.id, 16);
-      panel += `<a class="dark:text-indigo-400 dark:visited:text-indigo-400 dark:hover:text-indigo-500" href="https://meshview.armooo.net/packet_list/${nodeId}" target="_blank">Armooo's MeshView</a><br/>`;
-      panel += `<a class="dark:text-indigo-400 dark:visited:text-indigo-400 dark:hover:text-indigo-500" href="https://app.bayme.sh/node/${encodeURIComponent(node.id)}" target="_blank">Bay Mesh Explorer</a><br/>`;
-      panel += `<a class="dark:text-indigo-400 dark:visited:text-indigo-400 dark:hover:text-indigo-500" href="https://meshtastic.liamcottle.net/?node_id=${nodeId}" target="_blank">Liam's Map</a><br/>`;
-      panel += `<a class="dark:text-indigo-400 dark:visited:text-indigo-400 dark:hover:text-indigo-500" href="https://meshmap.net/#${nodeId}" target="_blank">MeshMap</a><br/>`;
-
-      nodeTitle.textContent = node.longname ?? "";
-      nodeSubtitle.textContent = node.shortname ?? "";
-      nodeContent.innerHTML = panel;
-      nodePanel.classList.remove("hidden");
+        neighborLayers.push(vectorLineLayer);
+        map.addLayer(vectorLineLayer);
+      });
     });
 
     // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -1451,307 +1070,29 @@ export function Map() {
   const usingMapbox = provider === "mapbox" && canUseMapbox;
 
   return (
-  <div className="relative w-full h-full min-h-0 overflow-hidden overscroll-none">
-    <div id="map" ref={mapRef} className="absolute inset-0" />
+    <div className="relative w-full h-full min-h-0 overflow-hidden overscroll-none">
+      <div id="map" ref={mapRef} className="absolute inset-0" />
 
-      {/* Bottom-right anchor: Legend + Map Settings stacked */}
-      <div className="fixed bottom-4 right-4 z-[1100]">
-        {/* Mobile toggle sits with the panel */}
-        <button
-          ref={settingsToggleRef}
-          type="button"
-          onClick={() => setSettingsPanelOpen((v) => !v)}
-          className="lg:hidden mb-2 ml-auto block p-2 rounded-lg shadow-lg backdrop-blur-sm border transition-all duration-200
-                    bg-white/90 dark:bg-gray-800/90 border-gray-200 dark:border-gray-600
-                    hover:bg-gray-50 dark:hover:bg-gray-700"
-          aria-label="Toggle Map Settings"
-          aria-expanded={settingsPanelOpen}
-        >
-          <div className="w-5 h-5 flex items-center justify-center">
-            {settingsPanelOpen ? (
-              <svg
-                className="w-4 h-4 text-gray-700 dark:text-gray-200"
-                fill="none"
-                stroke="currentColor"
-                viewBox="0 0 24 24"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M6 18L18 6M6 6l12 12"
-                />
-              </svg>
-            ) : (
-              <svg
-                className="w-4 h-4 text-gray-700 dark:text-gray-200"
-                fill="none"
-                stroke="currentColor"
-                viewBox="0 0 24 24"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M12 6V4m0 2a2 2 0 100 4m0-4a2 2 0 110 4m-6 8a2 2 0 100-4m0 4a2 2 0 100-4m0 4v2m0-6V4m6 6v10m6-2a2 2 0 100-4m0 4a2 2 0 100-4m0 4v2m0-6V4"
-                />
-              </svg>
-            )}
-          </div>
-        </button>
+      <MapSettingsPanel
+        settingsPanelRef={settingsPanelRef}
+        settingsToggleRef={settingsToggleRef}
+        settingsPanelOpen={settingsPanelOpen}
+        setSettingsPanelOpen={setSettingsPanelOpen}
+        provider={provider}
+        setProvider={setProvider}
+        mapboxStyle={mapboxStyle}
+        setMapboxStyle={setMapboxStyle}
+        osmBasemap={osmBasemap}
+        setOsmBasemap={setOsmBasemap}
+        recentDays={recentDays}
+        setRecentDays={setRecentDays}
+        clusterEnabled={clusterEnabled}
+        setClusterEnabled={setClusterEnabled}
+        canUseMapbox={canUseMapbox}
+        usingMapbox={usingMapbox}
+      />
 
-        {/* Map Settings (sits above legend) */}
-        <div
-          ref={settingsPanelRef}
-          className={`absolute right-0 bottom-full mb-2 w-64 max-w-[calc(100vw-2rem)] rounded-xl shadow-lg border border-gray-200/70 dark:border-gray-700/70 bg-white/95 dark:bg-gray-900/95 backdrop-blur-md transition-all duration-300 ease-in-out
-            ${
-              settingsPanelOpen
-                ? "translate-y-0 opacity-100 pointer-events-auto"
-                : "translate-y-2 opacity-0 pointer-events-none"
-            }
-            lg:translate-y-0 lg:opacity-100 lg:pointer-events-auto`}
-        >
-          <div className="p-4">
-            <div className="flex items-center justify-between mb-4">
-              <h3 className="font-semibold text-gray-900 dark:text-gray-100">
-                Map Settings
-              </h3>
-
-              {/* Close button for mobile */}
-              <button
-                type="button"
-                onClick={() => setSettingsPanelOpen(false)}
-                className="lg:hidden p-1 rounded-md hover:bg-gray-100 dark:hover:bg-gray-700 transition-colors"
-                aria-label="Close settings"
-              >
-                <svg
-                  className="w-4 h-4 text-gray-500 dark:text-gray-400"
-                  fill="none"
-                  stroke="currentColor"
-                  viewBox="0 0 24 24"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={2}
-                    d="M6 18L18 6M6 6l12 12"
-                  />
-                </svg>
-              </button>
-            </div>
-
-            <div className="space-y-4 text-sm">
-              {/* Provider */}
-              <div>
-                <label
-                  htmlFor="provider-select"
-                  className="text-xs font-medium uppercase tracking-wide text-gray-600 dark:text-gray-300 mb-2 block"
-                >
-                  Provider
-                </label>
-                <select
-                  id="provider-select"
-                  aria-label="Map provider selection"
-                  className="w-full rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-800 px-3 py-2 text-gray-900 dark:text-gray-100 focus:border-blue-500 dark:focus:border-blue-400 focus:outline-none focus:ring-1 focus:ring-blue-500 dark:focus:ring-blue-400"
-                  value={provider}
-                  onChange={(e) => setProvider(e.target.value as MapProvider)}
-                >
-                  <option value="osm">OSM (OpenLayers)</option>
-                  <option value="mapbox" disabled={!canUseMapbox}>
-                    Mapbox (GL JS)
-                    {!canUseMapbox ? " — token not configured" : ""}
-                  </option>
-                </select>
-              </div>
-
-              {/* Style/Basemap */}
-              {usingMapbox ? (
-                <div>
-                  <label
-                    htmlFor="mapbox-style-select"
-                    className="text-xs font-medium uppercase tracking-wide text-gray-600 dark:text-gray-300 mb-2 block"
-                  >
-                    Mapbox Style
-                  </label>
-                  <select
-                    id="mapbox-style-select"
-                    aria-label="Mapbox map style selection"
-                    className="w-full rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-800 px-3 py-2 text-gray-900 dark:text-gray-100 focus:border-blue-500 dark:focus:border-blue-400 focus:outline-none focus:ring-1 focus:ring-blue-500 dark:focus:ring-blue-400"
-                    value={mapboxStyle}
-                    onChange={(e) => setMapboxStyle(e.target.value)}
-                  >
-                    <option value="mapbox/dark-v11">Dark</option>
-                    <option value="mapbox/streets-v12">Streets</option>
-                    <option value="mapbox/satellite-streets-v12">
-                      Satellite Streets
-                    </option>
-                  </select>
-                </div>
-              ) : (
-                <div>
-                  <label
-                    htmlFor="osm-basemap-select"
-                    className="text-xs font-medium uppercase tracking-wide text-gray-600 dark:text-gray-300 mb-2 block"
-                  >
-                    OSM Basemap
-                  </label>
-                  <select
-                    id="osm-basemap-select"
-                    aria-label="OpenStreetMap basemap selection"
-                    className="w-full rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-800 px-3 py-2 text-gray-900 dark:text-gray-100 focus:border-blue-500 dark:focus:border-blue-400 focus:outline-none focus:ring-1 focus:ring-blue-500 dark:focus:ring-blue-400"
-                    value={osmBasemap}
-                    onChange={(e) =>
-                      setOsmBasemap(e.target.value as OsmBasemap)
-                    }
-                  >
-                    <option value="osm">OSM Standard</option>
-                    <option value="osm_hot">OSM HOT</option>
-                    <option value="carto_positron">Carto Positron (Light)</option>
-                    <option value="carto_dark">Carto Dark Matter (Dark)</option>
-                  </select>
-                </div>
-              )}
-
-              {/* Last Seen Filter */}
-              <div>
-                <label
-                  htmlFor="recent-days-select"
-                  className="text-xs font-medium uppercase tracking-wide text-gray-600 dark:text-gray-300 mb-2 block"
-                >
-                  Show Last Seen
-                </label>
-                <select
-                  id="recent-days-select"
-                  aria-label="Filter nodes by last seen timeframe"
-                  className="w-full rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-800 px-3 py-2 text-gray-900 dark:text-gray-100 focus:border-blue-500 dark:focus:border-blue-400 focus:outline-none focus:ring-1 focus:ring-blue-500 dark:focus:ring-blue-400"
-                  value={recentDays}
-                  onChange={(e) => setRecentDays(Number(e.target.value))}
-                >
-                  <option value={30}>30 days</option>
-                  <option value={14}>14 days</option>
-                  <option value={7}>7 days</option>
-                  <option value={5}>5 days</option>
-                  <option value={3}>3 days</option>
-                  <option value={1}>1 day</option>
-                </select>
-              </div>
-
-              {/* Clustering Toggle */}
-              <div className="flex items-center justify-between p-3 bg-gray-50 dark:bg-gray-800/50 rounded-lg">
-                <div className="flex flex-col">
-                  <label
-                    htmlFor="clustering-checkbox"
-                    className="text-sm font-medium text-gray-900 dark:text-gray-100"
-                  >
-                    Node Clustering
-                  </label>
-                  <p className="text-xs text-gray-500 dark:text-gray-400 mt-1">
-                    {!usingMapbox ? "Mapbox only" : "Group nearby nodes"}
-                  </p>
-                </div>
-                <div className="relative">
-                  <input
-                    id="clustering-checkbox"
-                    type="checkbox"
-                    checked={clusterEnabled}
-                    onChange={(e) => setClusterEnabled(e.target.checked)}
-                    disabled={!usingMapbox}
-                    className="h-4 w-4 rounded border-gray-300 text-blue-600 focus:ring-blue-500 disabled:opacity-50"
-                    aria-label="Toggle node clustering (Mapbox only)"
-                  />
-                </div>
-              </div>
-
-              {/* Info Note */}
-              {!canUseMapbox && (
-                <div className="text-xs text-gray-500 dark:text-gray-400 p-3 bg-gray-50 dark:bg-gray-800/50 rounded-lg">
-                  Mapbox features are disabled because{" "}
-                  <code className="bg-gray-200 dark:bg-gray-700 px-1 py-0.5 rounded text-xs">
-                    VITE_MAPBOX_TOKEN
-                  </code>{" "}
-                  is not configured.
-                </div>
-              )}
-            </div>
-          </div>
-        </div>
-
-        {/* Legend (anchored bottom-right) */}
-        <div
-          id="legend"
-          className="bg-white/95 dark:bg-gray-900/95 backdrop-blur-md rounded-lg shadow-lg border border-gray-200 dark:border-gray-700 p-3"
-        >
-          <div className="text-sm font-semibold text-gray-900 dark:text-gray-100 mb-2">
-            Legend
-          </div>
-          <div className="space-y-1 text-xs text-gray-600 dark:text-gray-300">
-            <div className="flex items-center gap-2">
-              <div className="w-4 h-1 bg-green-400 rounded-full" />
-              <span>Heard A Neighbor</span>
-            </div>
-            <div className="flex items-center gap-2">
-              <div className="w-4 h-1 bg-blue-400 rounded-full" />
-              <span>Heard By Neighbor</span>
-            </div>
-            <div className="flex items-center gap-2">
-              <div className="w-4 h-1 bg-purple-400 rounded-full" />
-              <span>Mutual Connection</span>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      {/* Node Details Panel */}
-      <div
-        id="details"
-        className="hidden fixed top-2 right-2 z-[1050]
-           w-[92vw] sm:w-80 max-w-[calc(100vw-1rem)]
-           bg-white dark:bg-gray-900 rounded-xl shadow-xl border border-gray-200 dark:border-gray-700
-           max-h-[60vh] sm:max-h-[calc(100vh-2rem)]
-           overflow-hidden flex flex-col"
-      >
-        <div className="flex items-center justify-between p-4 border-b border-gray-200 dark:border-gray-700">
-          <div className="flex-1 min-w-0">
-            <div
-              id="details-title"
-              className="text-lg font-semibold text-gray-900 dark:text-gray-100 truncate"
-            >
-              NODE NAME
-            </div>
-            <div
-              id="details-subtitle"
-              className="text-sm text-gray-500 dark:text-gray-400 truncate"
-            >
-              NODE
-            </div>
-          </div>
-          <button
-            onClick={() => {
-              clearMapboxSelectionAndOverlays();
-            }}
-            className="ml-3 p-1 rounded-md hover:bg-gray-100 dark:hover:bg-gray-700 transition-colors"
-            aria-label="Close details"
-          >
-            <svg
-              className="w-4 h-4 text-gray-500 dark:text-gray-400"
-              fill="none"
-              stroke="currentColor"
-              viewBox="0 0 24 24"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M6 18L18 6M6 6l12 12"
-              />
-            </svg>
-          </button>
-        </div>
-        <div
-          id="details-content"
-          className="p-4 overflow-y-auto min-h-0 flex-1 text-sm text-gray-700 dark:text-gray-300"
-        />
-      </div>
+      <MapDetailsPanel onClose={clearMapboxSelectionAndOverlays} />
 
       <style>
         {`
